+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add CSS file to have all components colors and another configs
 - Add Config file to override default configs (for example custom css)
 - Add simple CSS utilities helper and delete IgniterCss for this version
+- Add exclude components option for `mix mishka.ui.gen.components`
 
 ### Tests: